+++ kinto_signer/serializer.py
@@ -13,6 +13,8 @@ def canonical_json(records, last_modified):
 
     # Fix scientific notations of Python JSON to conform with ECMAScript v6
     # 9.30258908e-07 --> 9.30258908e-7
+    # https://www.ecma-international.org/ecma-262/6.0/#sec-tostring-applied-to-the-number-type
+    # ("no leading zero")
     dump = dump.replace("e-0", "e-")
 
     return dump